requirements.txt
numpy==1.16.4
Pillow==6.1.0
pystray==0.14.4
python-xlib==0.25
six==1.12.0
tornado==6.0.3
winregistry==0.8.3
